dev-requirements.txt
pytest-ckan
pytest-cov